Reorg index generation from upload
